Solve lilys homework in separate thread
Create LilysHomework class for the task. It implements interface Runnable,
so it's possilbe to run it in separate thread.
